+++ .github/workflows/ci.yml
@@ -115,7 +115,7 @@ jobs:
       - name: Generate new openapi.json
         run: |
           mkdir -p tmp
-          python scripts/generate_openapi_spec.py --output tmp/openapi.json --format json
+          python scripts/generate_openapi_spec.py --output tmp/openapi.json --format json --strip-modern-openapi-incompatibilities
       - name: Strip OpenAPI 3.1.0 incompatibilities from both files
         run: |
           python scripts/strip_openapi_incompatibilities.py openapi.json tmp/openapi_clean.json
