+++ scripts/strip_openapi_incompatibilities.py
@@ -9,24 +9,10 @@ import json
 import sys
 import os
 
-def strip_openapi_31_incompatibilities(spec):
-    """
-    Strip OpenAPI 3.1.0 specific fields that are incompatible with OpenAPI 3.0.x
-    This is needed for tools like oasdiff that only support OpenAPI 3.0.x
-    """
-    if isinstance(spec, dict):
-        # Remove exclusiveMinimum and exclusiveMaximum fields
-        spec.pop('exclusiveMinimum', None)
-        spec.pop('exclusiveMaximum', None)
-        
-        # Recursively process all values
-        for key, value in spec.items():
-            spec[key] = strip_openapi_31_incompatibilities(value)
-    elif isinstance(spec, list):
-        # Recursively process all items in lists
-        return [strip_openapi_31_incompatibilities(item) for item in spec]
-    
-    return spec
+# Add the scripts directory to the path so we can import openapi_utils
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+
+from openapi_utils import strip_openapi_31_incompatibilities
 
 def main():
     parser = argparse.ArgumentParser(description='Strip OpenAPI 3.1.0 incompatibilities from JSON file')
